# contracts/openapi/aura-v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aura Core API",
    "version": "1.5.0",
    "description": "Owner-authenticated agent, persona, conversation, run, and memory transport contract. Provider endpoints, credentials, and embedding vectors are never exposed by this API."
  },
  "servers": [{ "url": "/", "description": "Same-origin Aura BFF" }],
  "security": [{ "sessionCookie": [] }],
...
  "paths": {
...
    "/api/v1/agents": {
      "get": {
        "operationId": "listAgents",
        "tags": ["Agents"],
        "responses": {
          "200": { "description": "Owner-visible agent profiles and their current immutable revisions", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentCollection" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "operationId": "createAgent",
        "tags": ["Agents"],
        "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAgentRequest" } } } },
        "responses": {
          "201": { "description": "Agent profile and first immutable revision created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentDetail" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
    "/api/v1/agents/{agent_profile_id}": {
      "parameters": [{ "$ref": "#/components/parameters/AgentProfileId" }],
      "get": {
        "operationId": "getAgent",
        "tags": ["Agents"],
        "responses": {
          "200": { "description": "Agent profile and immutable revision history", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentDetail" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "operationId": "updateAgentStatus",
        "tags": ["Agents"],
        "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateAgentStatusRequest" } } } },
        "responses": {
          "200": { "description": "Agent lifecycle status updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentDetail" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
    "/api/v1/agents/{agent_profile_id}/revisions": {
      "post": {
        "operationId": "createAgentRevision",
        "tags": ["Agents"],
        "parameters": [{ "$ref": "#/components/parameters/AgentProfileId" }, { "$ref": "#/components/parameters/CsrfToken" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAgentRevisionRequest" } } } },
        "responses": {
          "201": { "description": "Immutable agent revision created and made current for the profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentDetail" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
...
    "/api/v1/agents/{agent_profile_id}/memory-policies": {
      "get": {
        "operationId": "listAgentMemoryPolicies",
        "tags": ["Agents", "Memories"],
        "parameters": [{ "$ref": "#/components/parameters/AgentProfileId" }],
        "responses": {
          "200": { "description": "Immutable owner-scoped memory policy revisions for the agent", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentMemoryPolicyCollection" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "post": {
        "operationId": "createAgentMemoryPolicy",
        "tags": ["Agents", "Memories"],
        "parameters": [{ "$ref": "#/components/parameters/AgentProfileId" }, { "$ref": "#/components/parameters/CsrfToken" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAgentMemoryPolicyRequest" } } } },
        "responses": {
          "201": { "description": "Validated immutable memory policy revision created without changing an agent revision", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentMemoryPolicy" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
    "/api/v1/agents/{agent_profile_id}/memory-policies/{policy_revision_id}/attach": {
      "post": {
        "operationId": "attachAgentMemoryPolicy",
        "tags": ["Agents", "Memories"],
        "parameters": [{ "$ref": "#/components/parameters/AgentProfileId" }, { "$ref": "#/components/parameters/MemoryPolicyRevisionId" }, { "$ref": "#/components/parameters/CsrfToken" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AttachAgentMemoryPolicyRequest" } } } },
        "responses": {
          "201": { "description": "New immutable agent revision created with the selected owner-scoped memory policy", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentDetail" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "__Host-aura_session", "description": "Opaque, rotated, Secure, HttpOnly, SameSite=Lax BFF session cookie." }
    },
    "parameters": {
      "AgentProfileId": { "name": "agent_profile_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
      "MemoryPolicyRevisionId": { "name": "policy_revision_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
      "CsrfToken": { "name": "X-CSRF-Token", "in": "header", "required": true, "schema": { "type": "string", "minLength": 32, "maxLength": 512 }, "description": "Session-bound anti-CSRF token returned by getSession." },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Caller-generated UUID. Reuse only for an exact replay of the same command." }
    },
    "responses": {
      "BadRequest": { "description": "Malformed request", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Unauthorized": { "description": "No valid Aura session", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Forbidden": { "description": "Authenticated subject is not the configured owner or command validation failed", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "NotFound": { "description": "Resource not found for this principal", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Conflict": { "description": "Version conflict, active run conflict, or idempotency-key payload conflict", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "UnprocessableEntity": { "description": "Semantically invalid command", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
    },
    "schemas": {
...
      "AgentMemoryPolicyCollection": {
        "type": "object", "additionalProperties": false,
        "required": ["items", "attachedPolicyRevisionId", "agentVersion"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/AgentMemoryPolicy" } },
          "attachedPolicyRevisionId": { "type": "string", "format": "uuid" },
          "agentVersion": { "type": "integer", "minimum": 1 }
        }
      },
      "AgentMemoryPolicy": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "agentProfileId", "revision", "sharedUserRead", "currentAgentRead", "sharedUserPromotion", "fallbackRelevanceThreshold", "maxMemories", "contextBudgetFraction", "fallbackAgentProfileIds", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "agentProfileId": { "type": "string", "format": "uuid" },
          "revision": { "type": "integer", "minimum": 1 },
          "sharedUserRead": { "type": "boolean" },
          "currentAgentRead": { "type": "boolean" },
          "sharedUserPromotion": { "type": "boolean" },
          "fallbackRelevanceThreshold": { "type": "number", "minimum": 0, "maximum": 1 },
          "maxMemories": { "type": "integer", "minimum": 1, "maximum": 8 },
          "contextBudgetFraction": { "type": "number", "exclusiveMinimum": 0, "maximum": 0.2 },
          "fallbackAgentProfileIds": { "type": "array", "maxItems": 64, "uniqueItems": true, "items": { "type": "string", "format": "uuid" }, "description": "Retrieval fallback grants only; these grants never authorize invocation or delegation." },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateAgentMemoryPolicyRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["sharedUserRead", "currentAgentRead", "sharedUserPromotion", "fallbackRelevanceThreshold", "maxMemories", "contextBudgetFraction", "fallbackAgentProfileIds", "expectedRevision"],
        "properties": {
          "sharedUserRead": { "type": "boolean" },
          "currentAgentRead": { "type": "boolean" },
          "sharedUserPromotion": { "type": "boolean" },
          "fallbackRelevanceThreshold": { "type": "number", "minimum": 0, "maximum": 1 },
          "maxMemories": { "type": "integer", "minimum": 1, "maximum": 8 },
          "contextBudgetFraction": { "type": "number", "exclusiveMinimum": 0, "maximum": 0.2 },
          "fallbackAgentProfileIds": { "type": "array", "maxItems": 64, "uniqueItems": true, "items": { "type": "string", "format": "uuid" } },
          "expectedRevision": { "type": "integer", "minimum": 1 }
        }
      },
      "AttachAgentMemoryPolicyRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["expectedAgentVersion"],
        "properties": { "expectedAgentVersion": { "type": "integer", "minimum": 1 } }
      },
...
      "ProfileStatus": { "type": "string", "enum": ["active", "disabled"] },
      "AgentCollection": {
        "type": "object", "additionalProperties": false,
        "required": ["items"],
        "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/AgentProfile" } } }
      },
      "AgentProfile": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "status", "version", "currentRevision", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "status": { "$ref": "#/components/schemas/ProfileStatus" },
          "version": { "type": "integer", "minimum": 1 },
          "currentRevision": { "$ref": "#/components/schemas/AgentRevision" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AgentDetail": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "status", "version", "currentRevision", "revisions", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "status": { "$ref": "#/components/schemas/ProfileStatus" },
          "version": { "type": "integer", "minimum": 1 },
          "currentRevision": { "$ref": "#/components/schemas/AgentRevision" },
          "revisions": { "type": "array", "items": { "$ref": "#/components/schemas/AgentRevision" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AgentRevision": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "profileId", "revision", "displayName", "purpose", "instructions", "personaRevisionId", "promptBundleRevisionId", "modelPolicyRevisionId", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "profileId": { "type": "string", "format": "uuid" },
          "revision": { "type": "integer", "minimum": 1 },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 255 },
          "purpose": { "type": "string", "minLength": 1, "maxLength": 2000 },
          "instructions": { "type": "string", "minLength": 1, "maxLength": 32768 },
          "personaRevisionId": { "type": "string", "format": "uuid" },
          "promptBundleRevisionId": { "type": "string", "format": "uuid" },
          "modelPolicyRevisionId": { "type": "string", "format": "uuid" },
          "memoryPolicyRevisionId": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateAgentRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["displayName", "purpose", "instructions", "personaRevisionId"],
        "properties": {
          "displayName": { "type": "string", "minLength": 1, "maxLength": 255 },
          "purpose": { "type": "string", "minLength": 1, "maxLength": 2000 },
          "instructions": { "type": "string", "minLength": 1, "maxLength": 32768 },
          "personaRevisionId": { "type": "string", "format": "uuid" }
        }
      },
      "CreateAgentRevisionRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["displayName", "purpose", "instructions", "personaRevisionId", "expectedVersion"],
        "properties": {
          "displayName": { "type": "string", "minLength": 1, "maxLength": 255 },
          "purpose": { "type": "string", "minLength": 1, "maxLength": 2000 },
          "instructions": { "type": "string", "minLength": 1, "maxLength": 32768 },
          "personaRevisionId": { "type": "string", "format": "uuid" },
          "expectedVersion": { "type": "integer", "minimum": 1, "description": "Expected current profile version." }
        }
      },
      "UpdateAgentStatusRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["status", "expectedVersion"],
        "properties": {
          "status": { "$ref": "#/components/schemas/ProfileStatus" },
          "expectedVersion": { "type": "integer", "minimum": 1, "description": "Expected current profile version." }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object", "additionalProperties": false,
        "required": ["type", "title", "status", "code", "detail", "retryable", "traceId", "resource"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_]{2,63}$" },
          "detail": { "type": "string", "minLength": 1, "maxLength": 1000 },
          "retryable": { "type": "boolean" },
          "traceId": { "type": "string", "pattern": "^[0-9a-f]{32}$" },
          "resource": { "type": ["string", "null"], "maxLength": 255 },
          "fieldErrors": { "type": "array", "items": { "$ref": "#/components/schemas/FieldError" } }
        }
      },
      "FieldError": {
        "type": "object", "additionalProperties": false,
        "required": ["field", "code", "message"],
        "properties": { "field": { "type": "string", "minLength": 1, "maxLength": 255 }, "code": { "type": "string", "minLength": 1, "maxLength": 64 }, "message": { "type": "string", "minLength": 1, "maxLength": 500 } }
      },
...
    }
  }
}